fix est cov and read length data

+++ preqc-lr/read_seq.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-void read_seq::set(string i, unsigned long long int l, double c)
+void read_seq::set(string i, unsigned long int l, double c)
 {
     read_id = i;
     read_len = l;
