+++ preqc-lr/read_seq.hpp
@@ -6,8 +6,8 @@ class read_seq
 {
   public:
     string read_id;
-    int read_len;
+    unsigned long int read_len;
     double cov;
-    void set(string i, unsigned long long int l, double c);
-    void updateCov( double c);
+    void set(string i, unsigned long int l, double c);
+    void updateCov(double c);
 };
